+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Version 0.0.8 (in development)
 - Added `negative?`, `positive?`, `max`, and `min` natives.
+- Added support for `binary!` literals. 
 
 
 # Version 0.0.7